Implement COI2 with loop.
